requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.7.4,<3.0.0
requests==2.31.0
google-generativeai>=0.7.0
langchain==0.2.16
langchain-google-genai==1.0.10
langchain-pinecone==0.1.3
langchain-community==0.2.16
langchain-huggingface==0.0.3
pinecone-client>=5.0.0,<6.0.0
streamlit==1.29.0
PyPDF2==3.0.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
typing-extensions==4.12.2
docx2txt==0.8